risk-register-import-export-tool/ai-service/requirements.txt
flask==3.0.3
flask-limiter==3.5.0
groq==0.9.0
python-dotenv==1.0.1
requests==2.31.0